chore: move variant vitest configs into test/config/
Reduce root clutter by relocating the 13 environment-specific vitest
configs (browser, bun, happy-dom, jsdom, react-server, vercel-edge,
workerd, and their smoke variants) into test/config/. vitest.config.ts
stays at the root since bare `vitest`/`pnpm test` and most editor
integrations auto-discover it there.

Also switches sharedConfig.globalSetup to import.meta.url-resolved
absolute paths, matching the alias entries already in the file. This
was needed for two reasons: knip's vitest plugin resolves globalSetup
string literals relative to the importing config file rather than
root (unlike vitest itself, which resolves relative to root), so it
started reporting false-positive unresolved imports once the variant
configs moved into a different directory. And moving the configs
under test/ pulled vitest.config.ts into the tsc program for the
first time (previously nothing under src/ or test/ referenced it),
surfacing a pre-existing noPropertyAccessFromIndexSignature error on
process.env.GITHUB_ACTIONS.
