+++ test/config/vitest.bun.smoke.config.ts
@@ -1,6 +1,6 @@
 import {defineConfig} from 'vitest/config'
 
-import {builtPackageAlias, smokeConfig} from './vitest.config'
+import {builtPackageAlias, smokeConfig} from '../../vitest.config'
 
 // Built-asset smoke test under the Bun runtime (run via `bun run vitest`), pinned to the
 // undici-backed Node entry that the `bun` condition resolves to (see test/exports.test.ts) -
