+++ test/config/vitest.happy-dom.smoke.config.ts
@@ -1,6 +1,6 @@
 import {defineConfig} from 'vitest/config'
 
-import {builtPackageAlias, smokeConfig} from './vitest.config'
+import {builtPackageAlias, smokeConfig} from '../../vitest.config'
 
 // Built-asset smoke test in a browser-like (happy-dom) environment, pinned to the fetch entry
 // browsers resolve to (see test/exports.test.ts).
