+++ test/config/vitest.jsdom.config.ts
@@ -1,11 +1,11 @@
 import {defineConfig} from 'vitest/config'
 
-import {nonNodeExclude, sharedConfig} from './vitest.config'
+import {nonNodeExclude, sharedConfig} from '../../vitest.config'
 
 export default defineConfig({
   test: {
     ...sharedConfig,
     exclude: nonNodeExclude,
-    environment: 'edge-runtime',
+    environment: 'jsdom',
   },
 })
