+++ test/config/vitest.react-server.config.ts
@@ -1,6 +1,6 @@
 import {defineConfig} from 'vitest/config'
 
-import {nonNodeExclude, sharedConfig} from './vitest.config'
+import {nonNodeExclude, sharedConfig} from '../../vitest.config'
 
 export default defineConfig({
   test: {
