+++ test/config/vitest.vercel-edge.config.ts
@@ -1,11 +1,11 @@
 import {defineConfig} from 'vitest/config'
 
-import {nonNodeExclude, sharedConfig} from './vitest.config'
+import {nonNodeExclude, sharedConfig} from '../../vitest.config'
 
 export default defineConfig({
   test: {
     ...sharedConfig,
     exclude: nonNodeExclude,
-    environment: 'jsdom',
+    environment: 'edge-runtime',
   },
 })
