+++ test/config/vitest.workerd.config.ts
@@ -1,7 +1,7 @@
 import {cloudflareTest} from '@cloudflare/vitest-pool-workers'
 import {defineConfig} from 'vitest/config'
 
-import {nonNodeExclude, sharedConfig} from './vitest.config'
+import {nonNodeExclude, sharedConfig} from '../../vitest.config'
 
 export default defineConfig({
   plugins: [
