+++ test/config/vitest.workerd.smoke.config.ts
@@ -1,7 +1,7 @@
 import {cloudflareTest} from '@cloudflare/vitest-pool-workers'
 import {defineConfig} from 'vitest/config'
 
-import {smokeConfig} from './vitest.config'
+import {smokeConfig} from '../../vitest.config'
 
 // Built-asset smoke test on workerd. The cloudflare pool resolves `get-it` through the real
 // `exports` map (no source alias) and, like real wrangler, excludes the `node` condition - so
